remove usermodel file
